added types file, moved all type defs there, modified hashtags component
