+++ frontend/src/api.ts
@@ -1,27 +1,8 @@
 import mapValues from 'lodash-es/mapValues'
-import { Tweet, keysToCamel } from './utils'
+import { keysToCamel } from './utils'
+import { GapResponse, SingleDateResponse, SummaryResponse } from './types'
 
 const API_URL = import.meta.env.VITE_API_URL
-
-type Calculation = {
-  tweet: number
-  retweet: number
-  retweetWithComment: number
-  time: number
-  hashtags: object
-}
-
-type SingleDateResponse = {
-  calculations: Calculation
-  tweets: Tweet[]
-}
-
-type GapResponse = {
-  longest_gap: number,
-  current_gap: number,
-}
-
-type SummaryResponse = Record<string, Calculation>
 
 export const fetchSingleDate = async (
   date: string
@@ -42,17 +23,17 @@ export const fetchSingleDate = async (
 }
 
 export const fetchGap = async (): Promise<GapResponse> => {
-  const response = await fetch(`${API_URL}running-gap`);
+  const response = await fetch(`${API_URL}running-gap`)
   if (response.status !== 200) {
     console.log(
       `Looks like there was a problem. Status Code: ${response.status}`
-    );
+    )
     return
   }
 
-  const responseData = await response.json();
-  return responseData;
-};
+  const responseData = await response.json()
+  return responseData
+}
 
 export const fetchSummary = async (): Promise<SummaryResponse> => {
   const response = await fetch(API_URL + 'summary')
